CloudFormation Multi-Cloud/lambda/requirements.txt
azure
boto3
google-api-python-client
requests